requirements.txt
Flask==3.0.3
opencv-python==4.10.0.84
numpy==1.26.4
PyYAML==6.0.1
SQLAlchemy==1.4.49
RPi.GPIO==0.7.1  # solo se usará en Raspberry Pi